Modularize Project.jsx
